chore(vue): Refactor render function based UI components to single-file components
