+++ packages/vue/src/components/index.ts
@@ -1,15 +1,13 @@
-export {
-  SignUp,
-  UserProfile,
-  UserButton,
-  OrganizationSwitcher,
-  OrganizationProfile,
-  CreateOrganization,
-  OrganizationList,
-  GoogleOneTap,
-  Waitlist,
-} from './uiComponents';
 export { default as SignIn } from './ui-components/SignIn.vue';
+export { default as SignUp } from './ui-components/SignUp.vue';
+export { default as GoogleOneTap } from './ui-components/GoogleOneTap.vue';
+export { default as Waitlist } from './ui-components/Waitlist.vue';
+export { default as OrganizationSwitcher } from './ui-components/OrganizationSwitcher.vue';
+export { default as CreateOrganization } from './ui-components/CreateOrganization.vue';
+export { default as OrganizationList } from './ui-components/OrganizationList.vue';
+export { UserProfile } from './ui-components/UserProfile';
+export { OrganizationProfile } from './ui-components/OrganizationProfile';
+export { UserButton } from './ui-components/UserButton';
 
 export {
   ClerkLoaded,
