+++ packages/vue/src/components/ui-components/GoogleOneTap.vue
@@ -1,0 +1,16 @@
+<script setup lang="ts">
+import { ClerkHostRenderer } from '../ClerkHostRenderer';
+import { useClerk } from '../../composables';
+import type { GoogleOneTapProps } from '@clerk/types';
+
+const clerk = useClerk();
+
+const props = defineProps<GoogleOneTapProps>();
+</script>
+
+<template>
+  <ClerkHostRenderer
+    :mount="() => clerk?.openGoogleOneTap(props)"
+    :unmount="clerk?.closeGoogleOneTap"
+  />
+</template>
